🚚: Split commands in to separate files
